WEAPON SYSTEM V1 FINISHED
Everything seems to be working fine, we have a R for manual reload and it's automatic as well when the player runs out of it. Also we have a simple UI (which will tell us the current amount of bullets + current weapon) that should be modified later on. switch between weapons with the nums 1 to 4 :)